--- assets/js/blocks/cart-checkout/cart-i2/columns/index.ts
@@ -1,1 +1,0 @@
-export * from './columns-block';
