--- assets/js/blocks/cart-checkout/checkout/columns/columns-block.tsx
@@ -1,15 +1,0 @@
-/**
- * External dependencies
- */
-import { useBlockProps } from '@wordpress/block-editor';
-
-export const Columns = ( {
-	children,
-	...props
-}: {
-	children?: React.ReactNode;
-} ): JSX.Element => {
-	const blockProps = useBlockProps( props );
-
-	return <div { ...blockProps }>{ children }</div>;
-};
