--- assets/js/blocks/cart-checkout/checkout/columns/index.ts
@@ -1,1 +1,0 @@
-export * from './columns-block';
