+++ assets/js/blocks/cart-checkout/checkout/edit.tsx
@@ -35,8 +35,7 @@ import type { TemplateArray } from '@wordpress/blocks';
  * Internal dependencies
  */
 import './styles/editor.scss';
-import { Columns } from './columns';
-import { addClassToBody, useBlockPropsWithLocking } from '../shared';
+import { addClassToBody, useBlockPropsWithLocking, Columns } from '../shared';
 import { CheckoutBlockContext, CheckoutBlockControlsContext } from './context';
 import type { Attributes } from './types';
