+++ assets/js/blocks/cart-checkout/shared/index.js
@@ -1,3 +1,4 @@
 export * from './hacks';
 export * from './use-forced-layout';
 export * from './editor-utils';
+export * from './columns-block';
